experiments/requirements.txt
pyyaml
networkx
torch==1.2
python-box==3.4.2
pylint==2.3.1
matplotlib==3.0.3
pytest==4.4.1
tensorboardX==1.8
numpy
tqdm
natsort==6.0.0
addict
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
torch-geometric
black==19.10b
flake8==3.7.9
graphlog

